+++ tests/testthat/test_benchmark.R
@@ -473,6 +473,13 @@ test_that("param_values in benchmark", {
   design = data.table(task = tasks, learner = learners, resampling = resamplings, param_values = list(list(list(cp = 0.1), list(minbucket = 2))))
   bmr = benchmark(design)
 
-  expect_equal(bmr$learners$learner[[1]]$param_set$values, list(xval = 0, minsplit = 12, minbucket = 2))
-  expect_equal(bmr$learners$learner[[2]]$param_set$values, list(xval = 0, minsplit = 12, cp = 0.1))
+  sortnames = function(x) {
+    if (!is.null(names(x))) {
+      x <- x[order(names(x))]
+    }
+    x
+  }
+
+  expect_equal(sortnames(bmr$learners$learner[[1]]$param_set$values), list(minbucket = 2, minsplit = 12, xval = 0))
+  expect_equal(sortnames(bmr$learners$learner[[2]]$param_set$values), list(cp = 0.1, minsplit = 12, xval = 0))
 })